Split the Environment in 2 parts to make it possible to refactor
The context contains metadata about the document, where the environment should
only be used for rendering context. By splitting those 2 purposes hope to get
more insights on how the parser behaves and why certain data is needed.

If we are able to clean the parser context from any external dependencies, we might be
able to store it with the document, or like some kind of meta entry, which can be loaded
on document rendering.