[FEAT] add support for raw html sanitizers
This feature introduces a new compiler pass that will allow us to
sanitize the html in a raw directive. We can disable the full raw
usage by setting `raw_node.escape` to `true` this will block all
html and displays the raw node as normal paragraph.

By default the raw directive is displayed as is and sanitized to
safe html, when raw is defined as html content

```
.. raw:: html

   <p>html here</p>
```

if the raw language is different than html we no nothing for now.

This pr also allows the users to define there own html sanitizer
from the configuration. Example can be found in the included test.